Fixed clean_leftovers function.

+++ inary/atomicoperations.py
@@ -341,11 +341,11 @@ class Install(AtomicOperation):
         # Chowning for additional files
         for _file in self.package.get_files().list:
             fpath = util.join_path(ctx.config.dest_dir(), _file.path)
-            if os.path.islink(fpath):
-                ctx.ui.info(_("Added symlink '{}' ").format(fpath), verbose=True)
-        #    else:
-        #        ctx.ui.info(_("Chowning in postinstall {0} ({1}:{2})").format(_file.path, _file.uid, _file.gid), verbose=True)
-        #        os.chown(fpath, int(_file.uid), int(_file.gid))
+            if os.path.islink(fpath)
+                if os.path.lexists(fpath) and os.path.exists(fpath):
+                    ctx.ui.info(_("Added symlink '{}' ").format(fpath), verbose=True)
+                else:
+                    ctx.ui.warning(_("Broken or missing symlink '{}'").format(fpath))
 
         if 'postOps' in self.metadata.package.isA:
             if ctx.config.get_option(
@@ -442,8 +442,8 @@ class Install(AtomicOperation):
                 old_file_path = os.path.join(
                     ctx.config.dest_dir(), old_file.path)
 
-                if old_file.path in new_paths:
-                    if os.path.islink(old_file_path):
+                if old_file.path not in new_paths:
+                    if os.path.islink(old_file_path) and os.path.lexists(old_file_path):
                         os.unlink(old_file_path)
                     continue
